Move more C++ source handling to OutFile
